[TAN-2343] Move as_geojson action to IdeaCustomFieldsController
