+++ back/engines/commercial/custom_maps/custom_maps.gemspec
@@ -18,6 +18,7 @@ Gem::Specification.new do |s|
   s.add_dependency 'ros-apartment', '>=2.9.0'
   s.add_dependency 'pundit', '~> 2.0'
   s.add_dependency 'rails', '~> 7.0'
+  s.add_dependency 'idea_custom_fields'
 
   s.add_development_dependency 'rspec_api_documentation'
   s.add_development_dependency 'rspec-rails'
