+++ back/engines/commercial/idea_custom_fields/app/policies/idea_custom_fields/idea_custom_field_policy.rb
@@ -22,3 +22,5 @@ module IdeaCustomFields
     end
   end
 end
+
+IdeaCustomFields::IdeaCustomFieldPolicy.prepend(CustomMaps::Patches::IdeaCustomFields::IdeaCustomFieldPolicy)
